requirements.txt
asgiref==3.7.2
boto3==1.34.20
botocore==1.34.20
dj-database-url==2.1.0
Django==5.0.1
django-crispy-forms==1.14.0
django-environ==0.11.2
django-storages==1.14.2
gunicorn==21.2.0
jmespath==1.0.1
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.0.7
